requirements.txt
tensorflow>=2.12.0
scikit-learn>=1.2.0
spacy>=3.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=1.5.0
numpy>=1.23.0
jupyter>=1.0.0
streamlit>=1.22.0
pillow>=9.4.0
opencv-python>=4.7.0
textblob>=0.17.1
joblib>=1.2.0
nltk>=3.7